third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-dy-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dy' property of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dropShadowElement.dy.baseVal is 0
PASS dropShadowElement.dy.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
